Added a new modules to facilitate selection and general utilities
naturalselection/naturalselection.fs:
Added a new module in which to define selection functions (which determines which members of a generation should be propagated to the next). Added a simple method which only permits the fittest members to survive.

utils/utility.fs:
Added a utility file for defining useful functions

organism/operations.fs:
Added function to produce a generation from a list of organisms